Use gate scheduler in dense_simulator
This also prints information in dense simulator.

We do not print the density and num_nonzeros because computing them slows down
the simulation considerably. This is due to the need to copy data back and
forth between the GPU. Perhaps we can calculate these values on the GPU side
and only pass the information to the simulator.

Moreover, we do not print num_gate_apps as this information is irrelevant for a
dense simulator.